Refactor EntCertEntitledContent.
This was a source of confusion when trying to write a content plugin as it's
very difficult to tell what object you're getting where, content is referenced
all over the place. In particular, this object carried a content, which callers
frequently referred to as content.content.

After some discussion, refactoring such that we leave rhsm.certificate2.Content
as a cert specific class, and use this as a more general class.

Removed the reference to the content object we're trying to hide, and copied
the data we use off it onto the new object. It continues to carry the
additional reference to it's cert.