1153375: Support TLSv1.2 and v1.1 by default.
To support servers that only support TLS v1.1 or
newer, the default TLS protocol needs to change.

Change default protocol name to "sslv23". "sslv23"
is an openssl ssl method for starting with TLS v1.2
and failing back to v1.1, v1, and SSL v3.

These changes are related to the "poodle" attack
(CVE-2014-3566) on SSLv3, and the push to move servers to
drop support for SSLv3 and to prefer TLSv1.2.

Note: With older versions of openssl, and servers that
provide SSLv3, this can allow the connection to negotiate
to SSLv3. Newer versions of openssl with support for the
TLS_FALLBACK_SCSV will prevent this.

Servers implementing the rhsm api (aka, candlepin) should
be updated to disable SSLv3. They will need to continue
to support TLSv1 for older client versions.