Remove ostree 'unconfigured' after configuring.
A ostree origin that has not been configured to
update it's remotes has a 'unconfigured-state' option
in the origin file. Once we set up remotes and update the
origin file, we remove that.

If it's not removed, 'atomic upgrade' will warn with the
message there ("Use "subscription-manager register" to enable online
updates" by default).

We try to only remove it on successful attaches.