1157387: Fix incorrect no installed products detected status in GUI.
Against SAM 1.4, it seems quite easy to end up with a compliance report that
has no compliant until date. This appears to be a separate bug already fixed in
master and hosted.

The client was incorrectly assuming that if status was valid, and we have no
compliant until date, we must not have any installed products. (normally this
would be a safe assumption)

Instead explicitly check if we have no installed products, and if not we'll
display a less precise message that omits the compliant until date. A warning
is logged as well.