1113741: Fix rhsmd traceback on 502 errors.
rhsmd invoked from cron with the '-s' option would traceback
if getCompliance encounters a 502 error and raising a
RemoteServerException. The try/except block in cache.py now
handles that exception as well.

Also combine the handler for NetworkException, socket.error,
and RemoteServerException. This also changed the cache
behavior on an empty cache with network errors so that
the network exception isn't propragated up. In some
odd cases, the result will just be a slightly less specific
error message.

Moved check for UnauthorizedException before it's
parent class RemoteServerException.

The caught exceptions here no longer get the traceback logged,
the exception msg is sufficient.