Move ostree config to /etc/ostree/remotes.d/redhat.conf
Split up the OstreeConfig usage, so we can track
what info came from the old location and clean it up.

If we see an old style config with remotes in it, we
delete them. And then write the remote config to the
new file.

Also clean up some heavy logging.