Break out of firstboot on network error.
If we return a failure code to the firstboot screens,
we end on the same screen again, with no option to skip
or to go back to a point we could skip. In practice this
means if a user selects to register in firstboot, and then
doesn't or can't complete the process, they are stuck in
firstboot.

This forces the rhsm_login module to report success in that
case, so firstboot advances past the screen, to the end, and
exits.