1191237: Allow firstboot to exit on network errors
If we want to exit out of firstboot, we have to return
success from the result of the 'Done' button, otherwise
we redisplay the screen.

We were returning familure status to the firstboot nav,
but now it just means to show our module again. And since
it is the first, last, and only firstboot module, 'Back'
or 'Next' aren't meaninful, so it was possible to get
stuck in firstboot if you weren't able to complete a
subscription. Either because of network errors (or
if we incorrectly get shown on systems with no working network)
or if there are server errors, or just no available accounts.

If we get exceptions that are normally handled by displaying
error dialogs, we go straight to the end of the finish
page instead. The initial server selections screen is a
bit of a special case, since it tries to fail softly on
network errors instead of raising exceptions. So we just
track a network_error flag on the widgets and check it
when deciding to precede or not.