1191237: Fix problems exitting firstboot on errors
In firstboot, if we encounter problems that may block
registration, return users to the "why register" info
page where there is an options to skip registering.

Previously it would return to either the login page
or stay on the current page in the face of those errors.
But in the context of firstboot, there was no escape
route to get out of firstboot.

Add hooks to registergui.RegisterScreen and
rhsm_login.moduleClass so they know there preffered
error_page. If error handlers are triggered, we
return to the info page on not fatal errors.