1191241: Handle network starting after subman does.
If subscription-manage-gui or firstboot starts before
the network is completely started (dhcp has created
/etc/resolv.conf, etc) then any attempt to use
socket.getaddrinfo() will fail and will continue
to fail for the life of the process. Unless we call
the glibc res_init() method to force a re-init of
the name resolution setup (reload resolv.conf
essentially) and purge the cache of the negative
name lookups.

So this adds a reset_resolver() method that the
registration gui and firstboot will call whenever
they go past the "Choose Server" screen.

Network errors in other parts of firstboot will
return the ui to the "Info Screen" preceding the
"Choose Server" screen, and therefore call
reset_resolver() as they progress.