1191237: Make proxy config "save" work in firstboot.
In firstboot, we were attempting to merge in any
existing RHN proxy configuration, but we were doing
it after the user provided proxy info in the network
configuration dialog and were overwritting it. For
the case of RHN proxy config existing, but with
proxyEnabled=0, we were unsetting the user provided
values and saving it.

This moves the RHN proxy info merging to before the
network config dialog, and also fixes the RHN proxy
info loading to not alter anything if it's disabled.