1195446: Only set global socket timeout on RHEL 5.
We were unintentionally setting it everywhere, with the
consequence of changing the default socket timeout for
any modules that import us.

For now, fix the version check. But also move it
out of module scope, and into connection related
objects. Add a module global timeout_altered to
track if we've set it already. Update it if
needed in the http connection _request calls.