1183382: Fix test case to work with dateutil 2.
There seems to be a bug in python-dateutil 2.0 or
higher in the tz.tzstr() method, where it fails if
the arg passed in is a regular string and not a
unicode string. This works around that by changing
the test to pass the tz string as a unicode string.

See https://bugzilla.redhat.com/show_bug.cgi?id=1197791
and https://bugs.launchpad.net/dateutil/+bug/1331576 for
the dateutil bug.