1133647: Fix messageWindow deprecation warning.
As part of a workaround for #865702, we were
accessing a gtk.MessageDialogs private label so we
could detect if it failed to parse the markup in
error messages from the server, which at the time
was the clonepin proxy. Access the private label
causes a DeprecationWarning.

Those messages have since been updated, and clonepin
decommissioned, so the workaround is no longer
needed.