1201727: Handle reasons with expired ent id
Reasons that include an 'entitlement_id' attribute that
references a entitlement id that is not currently in
cert_sorters valid_entitlement_certs (for ex, a
cert that has since expired) were causing tracebacks.

Add test cases to cover that scenario, and some additional
tests in test_reasons.py to improve coverage.