1211557: Fix crash when rsyslog not running.
Recent logging changes included a config for a presently unused syslog handler.
This however causes a full crash in yum plugin and subscription-manager when
trying to configure logging if rsyslog isn't running.

As this is meant to be an optional logging handler, remove it from the
configured handler and formatter keys.