1159163: Fix prod id del because of --disablerepo
If yum was invoked with --disablerepo=example-repo, and
example-repo is the only repo associated with a product id,
we would delete the product id.

The fix adds code to attempt to determine what yum repos
have been disabled temporarily. Temporarily in this case
meaning not via redhat.repo, but via the yum --disablerepo
command line option, or potentially via another yum plugin.

We load and parse redhat.repo ourself, and track repos that
are enabled there, and compare to the runtime list of enabled
repos yum computes (after it's cli options and config plugins
are applied). If anything is missing, we consider it temporarily
disabled.

When we check product certs looking for certs that may need to
be deleted, if that product is mapped to one of the temporarily
disabled repos, it is not considered for deletion now.

In addition, productid.ProductManager() was incorrectly checking
the version of 'yum'. Versions newer than 3.2.28 are required
since earlier versions don't track the repoid a package was
installed from. Yum version 3.4.x was not evaluating as greater
than 3.2.x, so product cert deletion wasn't attempted. This means
RHEL 7.x wasn't checking for product certs to remove, and was hiding
1159163. Now 7.x will attempt to check for product certs to delete
as well, and do it correctly. The RHEL 6.x case is fixed as well.