1199597: Fix UnicodeError from repolib's report
The str() for repolib.RepoActionReport would fail if
it tried to join together str's that could not be
decoded into ascii, when combining the translation for
'Added (new)' and a unicode string of the content label
or a repo section id. So make sure content info is
encoded to utf8 for display.

This bug is dependent on the absence of python-simplejson
on RHEL7. The 'ourjson' module will use either the built
in python 'json' module or simplejson if it is installed.
python-simplejson is not an rpm dep however, so it is
optionaly. However, the API of python's 'json' and
'simplejson' differ slightly. The built in 'json' module
will deserialize JSON strings into unicode strings allways,
where simplejson does not.

So if python-simplejson was installed, the repolib report
does not cause the UnicodeDecodeError, since the content
labels are str's. But without it (the default), and with
a LANG that provides translations for the report fields that
can not be encoded to 'ascii', you will see an error like:

'ascii' codec can't decode byte 0xc3 in position 8: ordinal not in
range(128)

So this changes the report code to handle any of the gettext
strings or content info being str's or unicode strs.

Test cases added for repolib and entcertlib Report classes.