Make rhsm-debug test cases clean up better.
We were creating temp test dirs but never
deleting them. So do that in the tearDown().

Use a tempdir via tempfile.mkdtemp for test temp
directory instead of os.getcwd()/testing-dir. tempdir
tmp directory is also likely on tmpfs, so a little
faster.

Add a final_destination_path attribute to rhsm-debugs
SystemCommand that tracks the name of the archive
that was written, so tests can clean up.

Add some test helper methods to remove some
duplication.