1180273: Migrate from RHN Classic without credentials
This commit builds on a previous effort to allow users to migrate from
RHN Classic without a password.  In this commit, the option for leaving
a profile on RHN Classic (it is the deletion of a profile that requires
a password) has been switched to a flag rather than an option as it was
previously.

Additionally this commit fixes an issue where users were not being
required to provide credentials to the destination when leaving a
profile on RHN Classic.