Cast to correct treestore column type
In previous pygtk versions, for tree store
models, we could set their type as 'int', and
anything that could be stringified would be
by pygtk, but that is no longer the case.

So for columns that are marked 'str', make sure
we pass them actual strings and not ints.

If you see errors from gtk tree* code like:
ValueError: Expected string or unicode but got 4<type 'int'>

The value for one of the columns is getting the wrong
type, and likely needs the value str()'ed. Or the
column type changed to something more accurate.