1149095: Fix error when yum updates subman modules
If product-id yum plugin was installed and enabled, and
a yum transaction updates subscription-manager code, we
could end up with the old version of 'injection', and the
new version of 'injectioninit'.

The injectioninit import was not at module scope, and would
not load the module until the plugin hook is called. But
the other subman modules in productids module scope would
cause 'injection' to be imported.

So yum starts up, loads our 'product-id' yum plugin module,
which loads 'injection' module into the process. Then we
update subscription-manager, replacing all of our modules
include 'injection' and 'injectioninit'. After the rpm
transaction runs, then we run the 'posttrans_hook' product
id plugin, which would then load 'injectioninit' from the
filesystem and get the new injectioninit.

If the new injection init references names that are in
the new 'injection' module, but not in the old one that
is already loaded into the yum process, we would get:

AttributeError: 'module' object has no attribute 'PROFILE_MANAGER"

Anything in the new injectioninit that needed the new injection
could fail.

If the subscription-manager yum plugin is enabled, it would
import 'injectioninit' in the 'postconfig_hook' step, before
the yum transaction changes anything installed. In that
scenario, the post transaction 'product-id' plugin would
get the old already loaded 'injectioninit' module, that matches
the old already loaded 'injection' module and avoid the
problem.