Add module ('gtk any') as Gtk ver abstraction
The main gui code now uses the same code for
gtk2 and gtk3, via the 'ga' abstraction module.

The 'ga' API looks just like the Gtk 3/gobject introspection
api. With the addition of a few wrapper methods and functions
to smooth imcompatiblities with Gtk2.

When targetting Gtk3, it just imports the modules.

When targetting Gtk2, it includes an assortment of enum
like classes to expose the flat namespace gtk.* constants
(gtk.CANCEL_OK, etc) as enums, ala Gtk3.

The gui code now imports out 'ga' module, and it acts
as a namespace in front of a very Gtk3 like API.

ie, Gtk3 code like:

  from gi.repository import Gtk
  label = Gtk.Label("foo label")

is now:

  from subscription_manager import ga
  lable = ga.Gtk.Label("foo label")

 And the gtk 2 code would have been:

  import gtk
  label = gtk.Label("foo label")