1224806: Prevent yum blocking on rhsm locks
Previously the rhsm yum plugin 'subscription-manager'
would block waiting for the rhsm cert lock to be
released. This file based lock is used by rhsm
code (subscription-manager, rhsmcert-worker, the
yum plugins) to try to make changes to entitlement
certs and redhat.repo mostly atomic, and to prevent
multiple processes from colliding (for example,
rhsmcertd-worker and the yum plugin, both of which
may attempt to update the redhat.repo at the same
time).

If an rhsm process has acquired the rhsm cert lock
when the yum plugin attempts to acquire it, the yum
plugin would wait for the lock to be released. If
the lock ended up being long lasting, this could
block the yum process until the lock is release.

Now the ActionInvoker classes allow a different
Locker implementation to be passed in. The yum plugin
code now includes YumRepoLocker that it uses to
implement a non-blocking lock acquire process. The
YumRepoLocker will not block to wait for a lock,
instead it just doesn't run the action.

For the yum plugin, this just means if it can't
acquire the rhsm cert lock without blocking, it
just doesn't attempt to update redhat.repo (or
update ent certs if cache_only=False). This prevents
yum from waiting to acquire the rhsm lock.

More tests for lock.py

Update lock.py classes to be new style classes.