Use find_lang --with-gnome for the gnome help
%find_lang will find the localized files for
the domain we tell it, but we use the 'rhsm'
gettext domain for the shared po/mo files.

But the gnome-help and omf files are localized
in the 'subscription-manager' gettext domain,
so run %find_lang for that domain as well, and
include the resulting subscription-manager.lang
in the %files macro as well.

find-langs finds omf files and gnome-help files
so this stops the rpm build warnings about
duplicate files.