Make desktop-file-validate warnings.
It was complaining about GenericName and Comment being
redundant, and indeed, they were. So they have been
removed.

'Encoding' type is also now deprecated and UTF-8 is
the default, so remove it as well.

The Catagories 'System' and 'Settings' are both
'Main' categories. And validate wants only one
main category. 'Settings' seems unused by anything
now, so we removed it.