rename, ga.Gtk -> ga_Gtk etc
Try importing from a 'ga' that is dynamic and imports
the gi.repository style names into the module. This
uses python's sys.metapath support to add a custom
importer to the "front" of sys.path. The custom
importer knows how to handle imports for
'subscription_manager.ga'. The importer is in the
ga_loader module, and is used to decide which impl
of 'ga' is used (gtk2 or gtk3).

Currently doing 'from subscription_manager.ga import GObject as ga_Gobject'
since for gtk2 cases, it's not really GObject, and I'd like
to not shadow well known names, but maybe it doesn't matter.

Add a ga_impls/ with gtk2 and gtk3 ga impl

The 'virtual' module 'ga' will import the correct
implementation from the ga_impls/ dir. ga_loader
decides which version to use. It currently defaults
to gtk3, with an env variable to override for testing.

Add ga_loader init to test/__init__.py

We need to setup the virtual modules before
anything starts importing them.

TODO: add a 'default_impl' module that packaging can
choose, so packaged versions of subman do not attempt
to decide on the ga impl at runtime.