test_cache would fail sometimes
TestEntitlementStatusCache.test_write_cache was trying
to wait for the write_cache thread to finish, but could
miss it if hadn't started yet.

Since we don't track the cache.py threads, just
try to load the expected file a few times, and if
it fails enough, fail the test. Differently klugey.