1240801: Use latest initial-setup API
When initial-setup and anaconda were rebased to more
recent versions, some of the API changed. Most notably
the move of the 'ui.gui.categories' package to 'ui.categories'.
rhsm_gui.py was updated to reflect that.

The required module layout also changed in a similar manner,
so Makefile and spec file also updated to reflect that.

This should resolve errors loading our initial-setup
module.