1136163: Ignore pythonpath to avoid selinux AVCs
For rhsmcertd-worker.py in particular, since it runs
as a child of rhsmcertd, which has very restrictive
selinux policy by default.

Because of either PYTHONPATH env var or system/site/user
changes to the pythonpath, rhsmcertd-worker.py could
end up attempting to find and open modules in path
that the selinux policy prevents it from accessing.

If the path ended up including '.', it would attempt
to read from cwd and cause denials.

So specify the '-E' and '-s' args to /usr/bin/python
-E     : ignore PYTHON* environment variables (such as PYTHONPATH)
-s     : don't add user site directory to sys.path; also
         PYTHONNOUSERSITE