1161157,1155954: Improve performance of Repository Dialog
Instead of immediately sending repo override changes to
the server on each checkbox click, queue them up and send
them when the 'Apply' button is clicked.

All communication with the server while saving/loading
data, is done async, and the user will be presented with
a progess window until the request completes.

Closing the dialog will detect if there were any local
changes made, and prompt the user to save changes before
closing.

The Remove All Overrides button will only be enabled if
there are currently overrides on the server for the
selected repository.

This commit also addresses varions layout/display issues
for GTK3.