1243260: Make proxy config dialog work.
The behaviour of widget.get_names() isn't neccasarily
the same name used for the widget/object 'id' in the
ui builder xml. And with Gtk.Builder, it usually
is not.

The proxy dialog init would try to set the state of the
dialog based on the widget names, but would fail because
of above mentioned change. So the state of the buttons
is incorrect.

To fix it, widgets.SubManBaseWidget's pull_widget() method
was changed to set the expected widget name, as well as
to elevate the widget name to the objects attributes.