1221273: Auto-attach failure should not short-circuit other parts of registration
The registration occurs and is successful regardless of auto-bind. If the serverurl is
set as per the bug, then the received identity cert is from an entitlement engine that
is no longer reflected in the system configuration. With this change the config is
updated to match the parameters of the registration.