1246680: Remove subscriptions from rhsm-debug
Older Candlepins don't have the call.
Candlepin 2.0 will not return subscriptions for less than the super admin level.
In the Candlepins where it is allowed, the default is no-subscriptions because the SOS report folks didn't really care to have them.

In the end, most of the data can be gleaned from the pool listing.