1242998, 1254550: Fix "already reg'ed" in initial-setup
Handle initial-setup running on a system that is potentially
already registered and/or attached to subscriptions better.

The RegisterWidget as used in initial-setup now detects
if the system is already registered. Current behavior is
to automatically move ahead to 'attach' stage of registration
if needed, or to the done screen if already registered and
attached.

RegisterWidget now runs choose_initial_screen as an idle_add
callback, so initial setup idle_add's it, then the callback
decides which screen to immedidetly move to from
initial_screen, and emit's signals to do it.

Refactor RegisterDialog so AutoBindWizard can subclass it easier.

Give AutoAttachWidget it's own choose_initial_screen
since it doesn't need to detect if a system is registered.

Populate gui with info from kickstart (mostly).

Add a Screen.populate() method in registergui for
populating the screens widgets with defaults or
the values from a kickstart.

Change _on_refresh_cb to accept an unused msg arg
for consistency. (1254550)

Add 'register-message' signal for non error info (for
example, displaying the info that a system is already
registered and attached).

Move to having RegisterWidget driven by RegInfo

Add a pre() to ChooseServerPage() that it can use to
decide if it should start there are go to attach, etc.

Some screen bookeeping simplified.

Add a RegisterWidget.current_screen python prop

Rearrange the methods in this class to make some sense

Cleanup, refactoring, docs, test fixes, etc.