884288: Make register widgets handle resizing.
For wide windows or dialogs, most widgets were set to
use a align mode of 'fill'. This causes things like entry
boxes to expand to the entire width of the parent. Combined
with widget package that was causing the RegisterWidget to
expand to cover the widgets register screen, many widgets
fill the entire width of the screen.

Make Entry widgets set a min and max width, but allow them
to expand and fill up to that max width.

Tweak most of the Labels to auto wrap. Enabling end ellipsing
on most labels, but also give them plenty of room to expand with
by setting the suggest number of lines before ellipsing. Labels
also get set to expand. That combo removes the behavior of
labels that require a ton of horizontal space, forcing the
RegisterWidget to be very wide (and causing entries and the
like to fill all of that space).

Add Gtk.Adjustments for any ScrolledWindows that need it. Remove
some Gtk.ViewPorts that are not needed (those holding a TreeView
for example).

Set min size requests for the ScrollWindows, so the defualt size
allocated doesn't hide the contents.

Replace some Gtk.Box based layouts with Gtk.Grid instead. Especially
those that need Gtk.Alignment widgets for formatting.

Generally reduce some of the spacing and margins of the widgets,
especially around the edges of notebook screens. It should be up
to the parent that holds the RegisterWidget to setup borders, spacing,
and margins.

Add size groups for some of the more fussier layouts.