1185958: Quieter ostree plugin sans ostree
If the ostree content plugin is installed on a system that
doesn't have ostree installed and/or it is not an active
ostree deployed system, more or less ignore ostree plugin
errors.

Add a new exception type for gi_wrapper related errors, and
give it a unthreatening str() to avoid cluttering logs with
generally harmless tracebacks.

It should be every uncommon for a user to get in this situation
unless they have specifially installed the ostree content plugin
on a system that does not use ostree.