1261006: Handle multiple nav button clicks
Before, if register gui nav button ('next', 'attach', etc)
was clicked and the next screens pre() was async, and the
nav button was clicked again before it was finished, the gui
would go ahead and skip to the next screen. If the double-next screen
depends on info set by the next pre(), this would fail in an
assortment of strange ways.

So now block the RegisterWidget 'proceed' signal until the
next screen is ready (currents Screen's 'ready' prop == True).

Complicating matters is the async nature of most Screens .pre()
(and the callback method passed to AsyncBackend.). Many of the
callbacks have many many exit points, so we needed a way to
indicate the bottom half of the callback was done. Screens
indicate the callback is done via Screen.pre_done().

The pre_done() needs to set Screen 'ready' prop == True.
In RegisterWidget the 'proceed' handler is blocked when
the screens 'ready' changes to False. When the 'ready'
propery changes back to True, the handler is unblocked.

RegisterWidget will also provide a 'screen-ready' widget
that follows the current screens 'ready' property.

RegisterDialog and RHSMSpoke have handlers for notify on
RegisterWidgets 'screen-ready' property. Since the nav buttons
live in the parent widgets, the 'screen-ready' property change
handlers update the nav buttons sensitivity.