1267034: Handle 401 with cert based auth
If we got here with an unexcepted RestlibException,
nothing would signal that the attach was finished
(in the sense that there is nothing else it can do).

In particular, 401's that have a valid candlepin
error response. Which could happen if you have the
CA certs for current serverurl, but have a valid consumer
cert from a different server, and you fail the consumer
cert auth check.