1262075,1267179: Fix back/cancel nav
Replace 'cancel' with stock 'back' button.

register_dialog.ui was using the stock 'cancel', but
rhsm_gui.ui was not. But change both to stock 'back'

Add a RegisterWidget.applied_screen_history that
tracks screens shown to the user, so that back
and error handling can do the right thing. This
uses UniqueList to keep only the last time a
screen was visited.

Make back button blocked/insens when in async
screen pre(). This avoid changing screens while
waiting for a thread to finish. It also blocks
the buttons until the screens are ready().

'back' now attempts to go back to the last gui
screen that was shown to the user.