1268028: Fix skipped auto attach in registergui
When using initial-setup to register a system, if the
'Manually attach subscriptions after registration' option
was selected in the credentials page, the gui would
start the select sla screen before starting the 'done'
process. So the done screen could be shown, then a
now paused progress bar, then eventually the results of
the select sla screens startup.

Fix is just to either finish, or start select sla, not
both.