1268095: Replace SLA radio buttons w/ combobox
The box of Gtk.RadioButton widgets on SelectSLAScreen
causes an assortment of odd behavior.

Most notably, it sometimes obscures some of the available
options by having them in offscreen areas of a scrollwindow.
With default themes, it is not very obvious there is an option
to scroll and see other radio buttons.

The scrolled window was originally added to handle cases where
there is a large number of options, and a Gtk.Box of radio buttons
would cause all sorts of layout issues. Especially when the
register widgets window is a small fixed size (RHEL6 firstboot
for example).

To alleviate some of those issues, replace the box of radio
buttons with a Gtk.ComboBox.