1264964: Ignore uuid=None on package sync
The 'package_upload' yum plugin from katello-agent
will call packageprofilelib._do_update with
consumer_uuid=None, resulting in an invalid request
to the rhsm REST API, if the consumer system is
not registered to rhsm.

So if we get as far as CacheManagers update_check with
consumer_uuid=None, just log it and return. We can't
raise an exception in that case without causing yum
to show the exception message on stderr for what is
a valid configuration.

For 'package_upload', it doesn't check if the consumer
if registered and calls into ActionClient.profilelib._do_update,
which is past any checks for registration. This change
just prevents it from making a request to the rhsm REST
API like 'PUT /consumers/None/packages'.

Rhsm registered systems with a valid consumer_uuid will
continue to make the package profile upload request.