1284120: Fix 'rules-version' error in 'version'
When running the cli 'version' command, the thread
for the server version check was racing the version
command output, and no default value for
'rules-version' was set. And 'version' was getting
a key error on self.server_versions.

Move server version thread stuff to it's own method

Use a threading.Event() to indicate the version check
is complete. After creating it, we set() it by default,
and let VersionCommand clear() it so it will wait() on
the event and block version command until we get the
answer from the server.

Create the server_version_event in the Cli base class,
so we can unset it in subclass inits (and before we
run _do_command).