976859: Only check server version if asked.
976859 is about the server version check done by
the cli being slow, which it is (at least for
invocations that otherwise don't use the network)

315d08e880 moved the check to use a thread so as
to not block, but 1284120 was a bug in it.

For simplicity, this commit removes the threading,
but also removes the server version check from all
cli sub commands except 'register' and 'version'
since it was of limited utility in practice.