Start moving fact collecting code into rhsm.facts
To allow the facts dbus service to load the facts, it
would be better if it didn't have to load up any
subman code or modules from /usr/share/rhsm.

This moves the most of the hardware and fact inventory
code into src/facts/. And starts installing them into
PYTHONSITELIB/rhsm/facts/.

This also splits hwpobe.py into two parts, one that
requires root level access (admin_facts.py) and the
rest of the hardware inventory in hwprobe.py.

The facts service currently requires running as root,
but it could actually be split into seperate object
paths or interfaces with polkit controlling access.
A further change would be to provide two services,
one running as root, one as a user.