Rename and reorg DMI under a firmware_info parent.
Even with just DMI as a firmware_info type, we
have to support not having, having it but having
dmidecode fail, and having it but using a dump file
as the sources of it's info for testing.

The new 'firmware_info' class has a factory method for
creating the correct firware info provider classes. At
the moment that means DMI, DMI from file, or a NullProvider.

At some point in the future, the bulk of DMI collection
will need to move back to a unpriv'ed service (users
can read most DMI from /sys on recent kernels).

firmware_info will also need to learn to support
additional firmware types (device-tree? ).

TODO: The facts modules are now essentially a library
module that will be exposed via a DBus service. So the
various gettext'ed print()'s scattered in there need
to go away.