Add DBus.Properties introspection support.
This is a little odd and incomplete at the moment,
but this adds a implementation of the
org.freedesktop.DBus.Introspectable.Introspect()
method to our base_services module/class.

This is to add support for generating the introspection
data for DBus.Properties.

Introspect() returns an xml string, so to add our properties,
we call the supers version, and then munge the XML tree
to add info about our properties. The property related
code is in base_properties.

We borrow some code from lvm-dubstep for this
(GPLv2, copyright Red Hat, Inc).

Currently this only supports properties of type string.