Add a Property() class to base_properties.
For a typical dbus property we need name,
value, value signature, and access perms. So
add a Property() class for that info.

Add constructor to BaseProperty that knows how to
build a string:string BaseProperties from a python dict.

For Facts services, add a Dbus.Property for a 'facts'
prop, whose value is a dictionary.