Some untested changes to cache expiration.
Add a expiration.py, with a Expiration object for
caches etc to use to keep track of start and expire times.

The file base caches should be creating Expirations for
themselves based on the filestamp and/or cache creation
time.

Remove some of my obnoxious use of functools.total_ordering
on collection (and the weird 'if collection > other_collection'
shenanigans).

Figuring out where/how to potentially support async cache
read/writes is a little odd so that is still wip.