Disable version.py generation via setup.py.
This commit makes some necessary corrections to the version.py
generation in setup.py, but then disables it.  I feel like the Makefile
works better for generating version.py since we need that file under
src/ in order to run the code directly from the source tree.
Nonetheless, if we do decide to use it, I want it to work correctly.