Remove sys.path shenanigans that break yum imports.
This code was originally added to augment the built-in Python XML parser
so that it would report line numbers.  Eventually, I determined that the
built-in parser wouldn't cut it and switched to LXML but forgot to
remove this code.  This snippet breaks yum because yum is explicitly
attempting to import the C implementation.