Do not endlessly add to the keys file.
Our xgettext extraction requires several passes due to our use of C,
Python, and desktop files, so we want xgettext to append to a single
file for each pass.  However, we want to start with a fresh file each
time so we don't end up with orphaned strings that languish in the keys
file forever.