Add a script to gather test modules in random order.
Use the script like `nosetests $(./random-order.py)`.  The script
accepts a single argument to use as the seed for the RNG so using the
same seed yields the same order.  There are also two options, --limit to
return no more than N results and --slice that can be used to shrink
the list of tests returned or select a specific range.