1259768: initial-setup: notify and block for async
In registergui, we now keep track of the threads that are spawned for
AsyncBackend, and AsyncBackend provides a method: block_until_complete
which blocks until all tracked threads are done.

To address initial-setup being closed before async actions are complete,
we now call the block_until_complete method in execute, meaning that
waiting for the threads to complete is part of the normal finish.

Additionally, we now notify the hub directly of any changes to the
registration status, so that if a user leaves the spoke early, the
status will be updated asynchronously.