1360427: Show error if browser is not detected
This prevents weird behavior, like having gedit show raw HTML. Instead,
we show an error with the URL, so that the user can do something useful
with it.

The browser detection has the potential to be outdated if a new browser
is released, but as long as the user has at least one of the coded
browsers present, it should respect their default browser, and
otherwise, they can copy the URL from the error dialog.