1382897: Don't always reenable register menu item
Because we use the same registration dialog for both the menu item and
the register button on the install tab, we need to account for the menu
item in the destroy handler. Prior to this change, the handler
unconditionally (re)-enabled whatever the widget that triggered the
dialog was, which is wrong for the register menu item.