1261215: Fix frozen progress bars
Fixed by moving the blocking calls to candlepin to separate threads. One
small downside to this is that more GUI updates happen asynchronously,
and thus this *might* be harder to test/verify. We can possibly fix that
without reverting this fix by having the registergui hang around a
little longer if necessary.

This is by no means comprehensive for fixing usage of HTTP calls on the
main thread in subscription-manager-gui...