1390258: Validate --remove-rhn-packages conflicting options
For both the rhn-migrate-classic-to-rhsm and sat5to6 tools, disallow
states where the system should continue to use rhn packages (i.e. states
where the system is still registered).

For RHN/RHSM, the only state is keep, where migration is performed, but
no changes are made to the legacy registration.

For the Satellite case, this is either unentitle, where the system is
still registered to the Satellite instance, and keep, where the system
is registered and consuming.