1404930: Provide GUI flow to fix proxy settings
If a proxy issue is detected, allow the user to change network settings
by presenting a dialog informing them of the issue. Additionally, this
change allows them to jump straight to the network config dialog by
selecting OK. Otherwise, we exit the GUI immediately.

It would be possible to enter the main window once the network settings
are fixed, but that was a much larger change.