Provide DBus objects for configuration, facts, and registration.
This commit creates DBus objects off the com.redhat.RHSM1 namespace.
Objects include Facts which is meant to gather all the relevant system
facts, Config which gives access to the subscription-manager
configuration settings, and RegisterServer which opens a domain socket
with another DBus object listening to allow for system registration.
The indirection over the domain socket is so that credentials will be
passed securely from one process to another instead of going over the
system bus.

This commit also retrofits the main subscription-manager code to use the
Facts and Config objects.