1423443: Handle IndexError during m2crypto request
Looks like this happens because the httplib code that handles
socket.error expects socket.error to be in a certain state. I'm guessing
this occurs when the socket.error raised is initialized with a string,
rather than a code and string.

Anyways, this will handle the case specially by having the m2crypto
wrapper detect IndexError explicitly, and raise a socket.error. Even
though there's a risk that the IndexError is raised internally for some
other reason, it seems harmless to simply map any IndexError encountered
in an m2crypto request to a socket.error.