1429657: Remove catch-all on register --force
At the time the catch-all was put into place, it seems it was being used
to handle the 410 Gone, but as of 5c48d059, it is handled in
`managerlib.unregister`, so the catch-all is now unnecessary, and
causing the register to proceed regardless of whether an unregister was
successful.

This change makes it so that `subscription-manager register --force`
behaves much more like `subscription manager unregister &&
subscription-manager register`, in that if there is a failure to
communicate with the server during the unregister, then the register
will not occur (we'll exit with an error code instead).

This prevents some cases where a consumer is registered twice (thus
effectively losing the original consumer, since it is deleted locally,
but remains server-side).