1435013: Add splay to all checks done by rhsmcertd
We get the seed for the random number generator from the getrandom
syscall (see 'man getrandom' for more info). The intervals for heal
and cert are maintained but we do not start those timers until the
completion of the initial delay. In this way if many systems are
started all at the same time, they will perform their cert and heal
checks at slightly different times. This was done in an attempt to
reduce peak load as a result restarting a large number of systems
simultaneously.

Add comment with link explaining why we need to generate random numbers this way

Do not delay (or splay) if given the -n option

Add max-splay-minutes cli option and maxSplayMinutes conf option

Ensure a number of minutes larger than RAND_MAX is not used for splay

Allow splay to be disabled by setting it to 0

Update default rhsm.conf and rhsmcertd man page for splay

Add new rhsmcertd option to completion script.